+++ frontend/src/Admin/Pages/Detailing/EditRealizationD/EditRealizationD.js
@@ -17,6 +17,8 @@ const EditRealizationD = () => {
   const user = localStorage.getItem("role");
 
   const handleDelete = (id) => {
+    if (!window.confirm("Czy na pewno chcesz usunąć tą realizacje?")) return;
+
     dispatch(deleteRealizationD(id));
   };
